Refactor to general util
I'm going to have a bunch of commands that I need to run, so I'm just
going to build a single executable to contain all of them. Not an ideal
setup, but don't want to manage several executables right now.

Currently the fixup code does not compile, so it is not used.